# Conflicts:
#	src/main/webapp/WEB-INF/applicationContext.xml
#	src/main/webapp/views/search.jsp

+++ src/main/java/com/epam/lowcost/controller/UserController.java
@@ -1,41 +1,67 @@
 package com.epam.lowcost.controller;
 
 import com.epam.lowcost.model.User;
-import com.epam.lowcost.service.UserService;
+import com.epam.lowcost.service.interfaces.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.ui.ModelMap;
+import org.springframework.web.bind.annotation.*;
 
 import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Map;
 
+import static com.epam.lowcost.util.EndPoints.*;
+
 @Controller
-@RequestMapping(value = "/user")
+@RequestMapping(value = USER)
+@SessionAttributes({"sessionUser", "number"})
 public class UserController {
 
     @Autowired
     UserService userService;
 
-    @GetMapping(value = "/all")
-    public String getAllUsers(Model model) {
-        model.addAttribute("users", userService.getAllUsers());
-        return "users";
+
+    @GetMapping(value = ALL + "/{pageId}")
+    public String getAllUsers(@PathVariable int pageId, @ModelAttribute(value = "sessionUser") User sessionUser, ModelMap model) {
+        if (!sessionUser.isAdmin()) {
+            return "redirect:" + TICKETS + SELF;
+        }
+
+        int usersByPage = (int) model.getOrDefault("number", 5);
+
+        Map<String, Object> pageRepresentation = userService.getUsersByPage(pageId, usersByPage);
+
+        model.addAttribute("pagesNum", pageRepresentation.get("pagesNum"));
+        model.addAttribute("users", pageRepresentation.get("users"));
+        model.addAttribute("pageId", pageRepresentation.get("pageId"));
+        return USERSPAGE;
     }
 
+    @GetMapping(value = SET_USERS_BY_PAGE)
+    public String setUsersByPage(@RequestParam String number, Model model) {
+        model.addAttribute("number", Integer.parseInt(number));
+        return "redirect:" + USER + ALL + FIRST_PAGE;
+    }
+
+
     @GetMapping
-    public String getById(@RequestParam long id, Model model) {
-        model.addAttribute("user", userService.getById(id));
+    public String getById(@ModelAttribute(value = "sessionUser") User sessionUser, @RequestParam long id, Model model) {
+        if (!sessionUser.isAdmin()) {
+            return "redirect:" + TICKETS + SELF;
+        }
+        List<User> user = new ArrayList<>();
+        user.add(userService.getById(id));
+        model.addAttribute("users", user);
         model.addAttribute("message", "Here is your User!");
-        return "users";
+        return USERSPAGE;
     }
 
     @PostMapping
     public String addUser(@RequestParam Map<String, String> params, Model model) {
-        model.addAttribute("user", userService.addUser(
+        User user = userService.addUser(
                 User.builder()
                         .email(params.get("email"))
                         .password(params.get("password"))
@@ -45,15 +71,14 @@ public class UserController {
                         .documentInfo(params.get("documentInfo"))
                         .birthday(LocalDateTime.parse(params.get("birthday")))
                         .isDeleted(false)
-                        .build()));
-
+                        .build());
+        model.addAttribute("user", user);
         model.addAttribute("message", "User successfully added");
-        return "users";
+        return USERSPAGE;
     }
 
-    @PostMapping(value = "/update")
+    @PostMapping(value = UPDATE)
     public String updateUser(@RequestParam Map<String, String> params, Model model) {
-
         User user = userService.updateUser(
                 User.builder()
                         .id(Long.valueOf(params.get("id")))
@@ -67,18 +92,69 @@ public class UserController {
                         .build());
         if (user == null) {
             model.addAttribute("message", "No such user or it has been deleted!");
+        }
+        if (params.get("userUpdate").equals("fromUser")) {
+            model.addAttribute("sessionUser", user);
+            return "redirect:" + TICKETS + SELF;
         } else {
             model.addAttribute("user", user);
             model.addAttribute("message", "User successfully updated");
         }
-        return "users";
+        return USERSPAGE;
     }
 
-    @PostMapping(value = "/delete")
-    public String deleteUser(@RequestParam long id, Model model) {
+    @PostMapping(value = ENROLL)
+    public String createUser(@RequestParam Map<String, String> params, Model model) {
+        userService.addUser(
+                User.builder()
+                        .email(params.get("email"))
+                        .password(params.get("password"))
+                        .isAdmin(Boolean.valueOf(params.get("isAdmin")))
+                        .firstName(params.get("firstName"))
+                        .lastName(params.get("lastName"))
+                        .documentInfo(params.get("documentInfo"))
+                        .birthday(LocalDateTime.parse(params.get("birthday")))
+                        .isDeleted(false)
+                        .build());
+        model.addAttribute("message", "Successfully registered. Please Log in. ");
+        return LOGIN;
+
+    }
+
+    @GetMapping(value = SETTINGS)
+    public String settings(@ModelAttribute("sessionUser") User sessionUser) {
+        return SETTINGSPAGE;
+    }
+
+    @PostMapping(value = CHANGE_PASSWORD)
+    public String changePassword(@ModelAttribute("sessionUser") User sessionUser, @RequestParam Map<String, String> params, Model model) {
+        User user = userService.verifyUser(sessionUser.getEmail(), params.get("oldPassword"));
+        if (user == null) {
+            model.addAttribute("message", "Wrong password");
+            return SETTINGSPAGE;
+        }
+        if (!params.get("newPassword").equals(params.get("newPassword2"))) {
+            model.addAttribute("message", "Passwords did not match!");
+            return SETTINGSPAGE;
+        }
+        sessionUser.setPassword(params.get("newPassword"));
+        userService.updateUser(sessionUser);
+        model.addAttribute("message", "Passwords changed successfully!");
+        return SETTINGSPAGE;
+    }
+
+    @PostMapping(value = DELETE)
+    public String deleteUser(@RequestParam long id, ModelMap model) {
+        User sessionUser = (User) model.get("sessionUser");
+        if (!sessionUser.isAdmin()) {
+            return "redirect:" + TICKETS + SELF;
+        }
+        if (sessionUser.getId() == id) {            
+            model.addAttribute("message", "You cant delete yourself!");
+            return "redirect:" + USER + ALL + FIRST_PAGE;
+        }
         model.addAttribute("message", userService.deleteUser(id));
-        return "users";
+        return "redirect:" + USER + ALL + FIRST_PAGE;
     }
-
 
 }
